+++ writer/writer_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/spf13/afero"
+	"github.com/verless/verless/config"
 	"github.com/verless/verless/fs"
 	"github.com/verless/verless/test"
 )
@@ -60,7 +61,7 @@ func TestWriter_removeOutDirIfExists(t *testing.T) {
 
 		testCase.beforeTest()
 
-		err := fs.Rmdir(memMapFs, w.outputDir)
+		err := fs.Rmdir(memMapFs, w.ctx.OutputDir)
 
 		if testCase.expectedError == "" {
 			test.Ok(t, err)
@@ -74,5 +75,11 @@ func TestWriter_removeOutDirIfExists(t *testing.T) {
 
 // setupNewWriter initializes a new writer instance.
 func setupNewWriter(fs afero.Fs) *writer {
-	return New(fs, testPath, testOutPath, false)
+	return New(Context{
+		Fs:                 fs,
+		Path:               testPath,
+		OutputDir:          testOutPath,
+		Theme:              config.DefaultTheme,
+		RecompileTemplates: false,
+	})
 }