+++ CHANGELOG.md
@@ -7,6 +7,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.4.0] - 2020-09-23
+
+### Added
+- Introduce the verless theming system.
+- Add the `verless create theme` command.
+
+### Changed
+- Create a default theme when creating a new project.
+- Update the example project to use the default theme.
+- Rename the `assets` directory to `static`.
+
+### Removed
+- Remove the `templates` directory.
+
+## [0.3.9] - 2020-09-22
+
+### Changed
+- Make entire page URL available as `Page.Href`.
+
+### Fixed
+- Fix wrong URLs for pages in the root directory.
+
+## [0.3.8] - 2020-09-22
+
+### Changed
+- Generate `list-page.html` even for empty routes.
+
 ## [0.3.7] - 2020-09-20
 
 ### Changed